+++ src/objects/webgl/index.ts
@@ -1,3 +1,4 @@
 export * from './Color';
 export * from './Program';
 export * from './Shader';
+export * from './Uniform';
